enh(interface): Add an ``allowed_entities`` setting in ``DerivativesDataSink``
Permit setting a list of ``allowed_entities`` at the instantiation time
of ``DerivativesDataSink``s so that it is possible to create arbitrary,
dynamic BIDS entities for the output file name without confusion with dynamic
metadata keys.

The reason to add this feature is to allow a more flexible interface for
derivatives generated via iterables in the context of
poldracklab/fmriprep#1558.