enh(reports): Refactor of the Report generation tools
This refactor is necessary in the context of poldracklab/fmriprep#1558
in order to allow a more flexible generation of reports.

The report factory now depends on a BIDSLayout, with a new configuration
file (by default, this is ``niworkflows/viz/figures.json``)

Ordering of tasks, runs, etc happens at query time.

The general structure of the reports is still handled by a
``config.json`` object, in which the BIDS entities to be queried are
listed (replacing the former regex patterns).

The overal Jinja template and the config file now look a bit cleaner
with several simplifications.