enh(interfaces): Add a ``KeySelect`` interface
This interface operates is akin to Python's ``OrderedDict``. It allows
to demultiplex list of dynamic lists (of same length) at the input
selecting the values corresponding to the requested key.

The implementation is proposed with several doctests ensuring the proper
functionality.

Pinging @satra and @effigies to know whether we should upstream this
interface directly to nipype.interfaces.utility