FIX: Segmentation plots aligned with cardinal axes
While nilearn/nilearn#25 is considered, this PR advances the generation
of mosaics after aligning the affine matrix of the data with the
cardinal axes. This should overcome the glitches when nilearn is
requested to plot contours and the mask for the ROI has voxels that only
partially intersect with the visualization plane.

Resolves: #542.
References: #281, #304.
Depends: #543.