FIX: Reportlets would crash in edge condition
The function that calculates the bounding box (and the cuts) to be
represented in reportlets has default thresholds thought out for brain
masks.

When passed a white-matter mask, these thresholds could yield only one
cut (below the required number of cuts).

This patch checkes that the number of cuts are sufficient in all the
cutting planes.

Resolves: #559.