enh: establish a documentation deployment protocol
This new protocol should give an end to the expensive builds on circle CI,
as old versions do not need to be rebuilt with every tag / push to master.

After squashing the history again, and probably cleaning up git objects that
are not used, this should effectively reduce the checkout time by a lot.

Resolves: #557.