fix(brain_extraction_wf): Pass inu_n4 to atropos_wf
antsBrainExtraction.sh passes the original N4 corrected image to
Atropos, where we have been running a masked N4 before Atropos.
We have examples where this difference leads to failures that do not
occur in the original ANTs workflow. Further, if we do run Atropos,
then this result is discarded and N4 is run using a white-matter mask.

This patch therefore simply passes the original N4 image to the Atropos
workflow, but otherwise leaves the workflow unchanged.